move envelopes in envelope.py.
